requirements.txt
openai>=1.0.0
twilio
python-dotenv
